refactor(cli): extract resetTerminalState to kilocode-specific file
Move the resetTerminalState helper out of the shared tui/util/terminal.ts
into packages/opencode/src/kilocode/cli/cmd/tui/util/terminal.ts to reduce
merge conflicts with upstream opencode.
